Split udSDK.cs into separate files and namespaces
