+++ languages/csharp/udSDK.cs
@@ -107,83 +107,6 @@ namespace Euclideon.udSDK
     }
   };
 
-  public enum RenderViewMatrix
-  {
-    Camera,     // The local to world-space transform of the camera (View is implicitly set as the inverse)
-    View,       // The view-space transform for the model (does not need to be set explicitly)
-    Projection, // The projection matrix (default is 60 degree LH)
-    Viewport,   // Viewport scaling matrix (default width and height of viewport)
-
-    Count,
-  };
-
-
-  public enum StandardAttribute
-  {
-    GPSTime,
-    ARGB,
-    Normal,
-    Intensity,
-    NIR,
-    ScanAngle,
-    PointSourceID,
-    Classification,
-    ReturnNumber,
-    NumberOfReturns,
-    ClassificationFlags,
-    ScannerChannel,
-    ScanDirection,
-    EdgeOfFlightLine,
-    ScanAngleRank,
-    LASUserData,
-
-    Count
-  };
-
-  public enum StandardAttributeContent
-  {
-    None = (0),
-    GPSTime = (1 << StandardAttribute.GPSTime),
-    ARGB = (1 << StandardAttribute.ARGB),
-    Normal = (1 << StandardAttribute.Normal),
-    Intensity = (1 << StandardAttribute.Intensity),
-    NIR = (1 << StandardAttribute.NIR),
-    ScanAngle = (1 << StandardAttribute.ScanAngle),
-    PointSourceID = (1 << StandardAttribute.PointSourceID),
-    Classification = (1 << StandardAttribute.Classification),
-    ReturnNumber = (1 << StandardAttribute.ReturnNumber),
-    NumberOfReturns = (1 << StandardAttribute.NumberOfReturns),
-    ClassificationFlags = (1 << StandardAttribute.ClassificationFlags),
-    ScannerChannel = (1 << StandardAttribute.ScannerChannel),
-    ScanDirection = (1 << StandardAttribute.ScanDirection),
-    EdgeOfFlightLine = (1 << StandardAttribute.EdgeOfFlightLine),
-    ScanAngleRank = (1 << StandardAttribute.ScanAngleRank),
-    LasUserData = (1 << StandardAttribute.LASUserData),
-  };
-
-  [StructLayout(LayoutKind.Sequential)]
-  public struct udRenderInstance
-  {
-    public IntPtr pointCloud;
-    [MarshalAs(UnmanagedType.ByValArray, SizeConst=16)]
-    public double[] worldMatrix;
-
-    public IntPtr filter;
-    public IntPtr voxelShader;
-    public IntPtr voxelUserData;
-    public double opacity;
-    public uint skipRender;
-  }
-
-  [StructLayout(LayoutKind.Sequential)]
-  public struct udAttributeSet
-  {
-    public StandardAttributeContent standardContent; //!< Which standard attributes are available (used to optimize lookups internally), they still appear in the descriptors
-    public uint count; //!< How many udAttributeDescriptor objects are used in pDescriptors
-    public uint allocated; //!< How many udAttributeDescriptor objects are allocated to be used in pDescriptors
-    public IntPtr pDescriptors; //!< this contains the actual information on the attributes
-  };
-
   [StructLayout(LayoutKind.Sequential)]
   public struct udPointCloudHeader
   {
@@ -194,7 +117,7 @@ namespace Euclideon.udSDK
     [MarshalAs(UnmanagedType.ByValArray, SizeConst = 16)]
     public double[] storedMatrix; //!< This matrix is the 'default' internal matrix to go from a unit cube to the full size
 
-    public udAttributeSet attributes;   //!< The attributes contained in this pointcloud
+    public Attributes.udAttributeSet attributes;   //!< The attributes contained in this pointcloud
 
     [MarshalAs(UnmanagedType.ByValArray, SizeConst = 3)]
     public double[] baseOffset; //!< The offset to the root of the pointcloud in unit cube space
@@ -300,115 +223,6 @@ namespace Euclideon.udSDK
     private static extern udError udContext_Disconnect(ref IntPtr ppContext, int endSession);
   }
 
-  public class udRenderContext
-  {
-    public void Create(udContext context)
-    {
-      udError error = udRenderContext_Create(context.pContext, ref pRenderer);
-      if (error != udError.udE_Success)
-        throw new UDException(error);
-
-      this.context = context;
-    }
-
-    public void Destroy()
-    {
-      udError error = udRenderContext_Destroy(ref pRenderer);
-      if (error != udError.udE_Success)
-        throw new UDException(error);
-    }
-
-    public void Render(udRenderTarget renderView, udRenderInstance[] pModels, int modelCount)
-    {
-      udError error = udRenderContext_Render(pRenderer, renderView.pRenderTarget, pModels, modelCount, (IntPtr)0);
-      if (error != udError.udE_Success)
-        throw new UDException(error);
-    }
-
-    public IntPtr pRenderer = IntPtr.Zero;
-    private udContext context;
-
-    [DllImport("udSDK")]
-    private static extern udError udRenderContext_Create(IntPtr pContext, ref IntPtr ppRenderer);
-    [DllImport("udSDK")]
-    private static extern udError udRenderContext_Destroy(ref IntPtr ppRenderer);
-    [DllImport("udSDK")]
-    private static extern udError udRenderContext_Render(IntPtr pRenderer, IntPtr pRenderView, udRenderInstance[] pModels, int modelCount, IntPtr options);
-  }
-
-  public class udRenderTarget
-  {
-    public void Create(udContext context, udRenderContext renderer, UInt32 width, UInt32 height)
-    {
-      udError error = udRenderTarget_Create(context.pContext, ref pRenderTarget, renderer.pRenderer, width, height);
-      if (error != udError.udE_Success)
-        throw new UDException(error);
-
-      this.context = context;
-    }
-
-    public void Destroy()
-    {
-      if (colorBufferHandle.IsAllocated)
-        colorBufferHandle.Free();
-      if (depthBufferHandle.IsAllocated)
-        depthBufferHandle.Free();
-
-      udError error = udRenderTarget_Destroy(ref pRenderTarget);
-      if (error != udError.udE_Success)
-        throw new UDException(error);
-    }
-
-    public void SetTargets(ref UInt32[] colorBuffer, UInt32 clearColor, ref float[] depthBuffer)
-    {
-      if (colorBufferHandle.IsAllocated)
-        colorBufferHandle.Free();
-      if (depthBufferHandle.IsAllocated)
-        depthBufferHandle.Free();
-
-      colorBufferHandle = GCHandle.Alloc(colorBuffer, GCHandleType.Pinned);
-      depthBufferHandle = GCHandle.Alloc(depthBuffer, GCHandleType.Pinned);
-
-      udError error = udRenderTarget_SetTargets(pRenderTarget, colorBufferHandle.AddrOfPinnedObject(), clearColor, depthBufferHandle.AddrOfPinnedObject());
-      if (error != udError.udE_Success)
-        throw new UDException(error);
-    }
-
-    public void GetMatrix(RenderViewMatrix matrixType, double[] cameraMatrix)
-    {
-      udError error = udRenderTarget_GetMatrix(pRenderTarget, matrixType, cameraMatrix);
-      if (error != udError.udE_Success)
-        throw new UDException(error);
-    }
-
-    public void SetMatrix(RenderViewMatrix matrixType, double[] cameraMatrix)
-    {
-      udError error = udRenderTarget_SetMatrix(pRenderTarget, matrixType, cameraMatrix);
-      if (error != udError.udE_Success)
-        throw new UDException(error);
-    }
-
-    public IntPtr pRenderTarget = IntPtr.Zero;
-    private udContext context;
-
-    private GCHandle colorBufferHandle;
-    private GCHandle depthBufferHandle;
-
-    [DllImport("udSDK")]
-    private static extern udError udRenderTarget_Create(IntPtr pContext, ref IntPtr ppRenderView, IntPtr pRenderer, UInt32 width, UInt32 height);
-
-    [DllImport("udSDK")]
-    private static extern udError udRenderTarget_Destroy(ref IntPtr ppRenderView);
-
-    [DllImport("udSDK")]
-    private static extern udError udRenderTarget_SetTargets(IntPtr pRenderView, IntPtr pColorBuffer, UInt32 colorClearValue, IntPtr pDepthBuffer);
-
-    [DllImport("udSDK")]
-    private static extern udError udRenderTarget_GetMatrix(IntPtr pRenderView, RenderViewMatrix matrixType, double[] cameraMatrix);
-
-    [DllImport("udSDK")]
-    private static extern udError udRenderTarget_SetMatrix(IntPtr pRenderView, RenderViewMatrix matrixType, double[] cameraMatrix);
-  }
 
   public class udPointCloud
   {
@@ -448,61 +262,6 @@ namespace Euclideon.udSDK
     private static extern udError udPointCloud_GetMetadata(IntPtr pModel, ref string ppJSONMetadata);
   }
 
-  public class udConvertContext
-  {
-    public void Create(udContext context)
-    {
-      udError error = udConvert_CreateContext(context.pContext, ref pConvertContext);
-      if (error != udError.udE_Success)
-        throw new UDException(error);
-    }
-
-    public void Destroy()
-    {
-      udError error = udConvert_DestroyContext(ref pConvertContext);
-      if (error != udError.udE_Success)
-        throw new UDException(error);
-    }
-
-
-    public void AddFile(string fileName)
-    {
-      udError error = udConvert_AddItem(pConvertContext, fileName);
-      if (error != udError.udE_Success)
-        throw new UDException(error);
-    }
-    public void SetFileName(string fileName)
-    {
-      udError error = udConvert_SetOutputFilename(pConvertContext, fileName);
-      if (error != udError.udE_Success)
-        throw new UDException(error);
-    }
-
-    public void DoConvert()
-    {
-      udError error = udConvert_DoConvert(pConvertContext);
-      if (error != udError.udE_Success)
-        throw new UDException(error);
-    }
-
-    public IntPtr pConvertContext;
-
-    [DllImport("udSDK")]
-    private static extern udError udConvert_CreateContext(IntPtr pContext, ref IntPtr ppConvertContext);
-
-    [DllImport("udSDK")]
-    private static extern udError udConvert_DestroyContext(ref IntPtr ppConvertContext);
-
-    [DllImport("udSDK")]
-    private static extern udError udConvert_AddItem(IntPtr pConvertContext, string fileName);
-
-    [DllImport("udSDK")]
-    private static extern udError udConvert_SetOutputFilename(IntPtr pConvertContext, string fileName);
-
-    [DllImport("udSDK")]
-    private static extern udError udConvert_DoConvert(IntPtr pConvertContext);
-
-  }
 
   public class udServerAPI
   {
